+++ src/components/TodoList/TodoList.tsx
@@ -1,1 +1,10 @@
-export const TodoList = () => {};
+import { Todo } from '../../types/Todo';
+import { TodoInfo } from '../TodoInfo';
+
+type Props = {
+  todos: Todo[];
+};
+
+export const TodoList: React.FC<Props> = ({ todos }) => {
+  return todos.map(todo => <TodoInfo key={todo.id} todo={todo} />);
+};
